requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
cycler==0.12.1
decorator==5.1.1
editdistance==0.8.1
flatbuffers==24.3.25
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.1
h5py==3.11.0
idna==3.7
imageio==2.34.2
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.6.0
requests==2.32.3
rich==13.7.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==71.1.0
six==1.16.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tkthread==0.5.2
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
